+++ haddock/haddock3.py
